계정 설정 완료

+++ src/main/java/com/board/weare/service/AccountServiceImpl.java
@@ -2,7 +2,6 @@ package com.board.weare.service;
 
 import com.board.weare.config.jwt.JwtProvider;
 import com.board.weare.dto.AccountDto;
-import com.board.weare.dto.CustomUserDetails;
 import com.board.weare.entity.Account;
 import com.board.weare.exception.BadRequestException;
 import com.board.weare.exception.MyEntityNotFoundException;
@@ -13,8 +12,6 @@ import com.board.weare.service.interfaces.AccountService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -40,17 +37,18 @@ public class AccountServiceImpl implements AccountService {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         Object principal = authentication.getPrincipal();
         log.info(principal.toString());
-        if(principal.equals("anonymousUser"))throw new UnauthorizedException("계정이 확인되지 않습니다. 다시 로그인 해주세요.");
+        if (principal.equals("anonymousUser")) throw new UnauthorizedException("계정이 확인되지 않습니다. 다시 로그인 해주세요.");
         return (Account) principal;
     }
 
     @Transactional
-    public Account login(AccountDto.Login loginDto){
+    public Account login(AccountDto.Login loginDto) {
         String username = loginDto.getUsername();
         String password = loginDto.getPassword();
         if (username == null || password == null) throw new BadRequestException("아이디/비밀번호는 필수입니다.");
-
-        jwtProvider.
+        Account account = get(username);
+        if(bCryptPasswordEncoder.matches(loginDto.getPassword(), account.getPassword()))return account;
+        throw new UnauthorizedException("로그인에 실패했습니다.");
 
     }
 
@@ -61,32 +59,11 @@ public class AccountServiceImpl implements AccountService {
         throw new MyEntityNotFoundException("account", "계정을 찾을 수 없습니다.");
     }
 
-    @Transactional(readOnly = true)
-    public List<Account> getAll() {
-        return accountRepository.findAll();
-    }
-
-    public Account getByTokensWhenNotExpired(String accessToken, String refreshToken) {
-        Optional<Account> accountOptional = accountRepository.findByAccessTokenAndRefreshToken(accessToken, refreshToken);
-        if (accountOptional.isPresent()) {
-            Account account = accountOptional.get();
-            Long accessTokenExp = account.getAccessTokenExp();
-            Long refreshTokenExp = account.getRefreshTokenExp();
-            Long now = new Date().getTime();
-            if (accessTokenExp < now && refreshTokenExp < now) throw new TokenExpiredException();
-            //TODO 하나라도 만료되면 재발급 과정 있어야 할듯?
-            return account;
-        }
-//        logger.debug("토큰체킹실패");
-        throw new UnauthorizedException();
-    }
-
-
     @Transactional
     public Account postAccount(AccountDto.Post req) {
         String username = req.getUsername();
         String password = req.getPassword();
-        String bcrypPassword =bCryptPasswordEncoder.encode(password);
+        String bcrypPassword = bCryptPasswordEncoder.encode(password);
         String name = req.getName();
         String phone = req.getPhone();
         LocalDateTime birthday = req.getBirthday();
@@ -102,9 +79,10 @@ public class AccountServiceImpl implements AccountService {
 
 
     @Transactional
-    public void patch(AccountDto.UpdateInfo updateInfo) {
-        String accountId = updateInfo.getId();
-        Account account = get(accountId);
+    public Account patch(AccountDto.UpdateInfo updateInfo) {
+        Account myInfo = getAccountFromSecurityContext();
+        String username = myInfo.getUsername();
+        Account account = get(username);
 
         String name = updateInfo.getName();
         String phone = updateInfo.getPhone();
@@ -114,34 +92,26 @@ public class AccountServiceImpl implements AccountService {
         String addressDetail = updateInfo.getAddressDetail();
         String zipcode = updateInfo.getZipcode();
         LocalDateTime birthday = updateInfo.getBirthday();
-        account.updateInfo(name, role, phone,email, address, addressDetail, zipcode, birthday);
+        account.updateInfo(name, role, phone, email, address, addressDetail, zipcode, birthday);
 
         //  비밀번호를 바꿀건가요?
         String password = updateInfo.getPassword();
+        if (password != null) {
+            account.updatePassword(bCryptPasswordEncoder.encode(password));
+        }
+
+        return account;
     }
 
-    @Transactional
-    public void patch(String accountId, String name, String role, String phone, String email, String address, String addressDetail, String zipcode, LocalDateTime birthday) {
-        Account account = get(accountId);
-
-        log.info("계정 수정");
-
-        // IDM 서버에 직원 계정 수정.
-        log.debug("IDM서버 직원 수정 전");
-        idmApi.patchAccount(accountId, name, role);
-
-        // AUTH 서버에 직원 계정 수정.
-        log.debug("Auth서버 직원 수정 전");
-        authApi.patchAccount(accountId, name);
-
-        // MCALL 서버 직원 계정 수정.
-        log.debug("MCall서버 직원 수정 전");
-        account.updateInfo(name, role, phone, email, address, addressDetail, zipcode, birthday);
-    }
-
-    public void delete(String id) {
+    public boolean delete(String id) {
         Account account = get(id);
-        accountRepository.delete(account);
+        Account myInfo = getAccountFromSecurityContext();
+
+        if (myInfo.isRoot() || account.getUsername().equals(id)) {
+            accountRepository.delete(account);
+            return true;
+        }
+        return false;
     }
 
 }
